Completed for lab5

+++ lab5lib/rectangle.h
@@ -4,7 +4,7 @@
 
 class Rectangle : public Shape {
 public:
-	Rectangle(Point2d pos, unsigned int* colorAlpha, float height, float width);
+	Rectangle(Point2d pos, uint8_t* colorAlpha, float height, float width);
 
 	void render(SDL_Renderer* renderer);
 private:
