+++ lab5lib/triangle.h
@@ -5,7 +5,7 @@
 class Triangle : public Shape {
 public:
 
-	Triangle(Point2d pos, unsigned int* colorAlpha, float base, float height);
+	Triangle(Point2d pos, uint8_t* colorAlpha, float base, float height);
 
 	void render(SDL_Renderer* renderer);
 private:
